backend/requirements.txt
Django==4.2.13
pillow
djangorestframework
setuptools
django-extra-fields
psycopg2